corrige formatação padrão da função McDate.getYear (refs #1880)

+++ src/protected/application/lib/modules/Components/assets/js/components-base/McDate.js
@@ -54,7 +54,7 @@ class McDate {
     }
 
     year(format) {
-        format = format || 'long';
+        format = format || 'numeric';
 
         return this.format({year: format});
     }
